requirements.txt
Flask==2.3.3
Flask-PyMongo==2.3.0
flask-login==0.6.3
flask-socketio==5.3.6
eventlet==0.33.3
pymongo==4.6.1
redis==5.0.1
requests==2.31.0
PyJWT[crypto]==2.8.0
dnspython==2.6.1
